chore(web): cleanup and consolidation
